Create FrontMatter.md
Created a markdown of just the top sections of each document. This will eventually be combined with the 36 different PnPs for each family. The purpose is to allow us to focus on just the family information in each of the other files and centralize any future changes to a single source.